Refactor BlockStreamService to extract BlockAccessService
